Add a blog section
The blog itself is a work in progress, there are no links to the blog
from the rest of the website and robots.txt will prevent the blog pages
from being indexed by search engines.

The framework used to enable blogging was adapted from the
jekyll-bootstrap project: https://github.com/plusjade/jekyll-bootstrap

Disqus is used to host comments/discussions.